requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.0
python-multipart>=0.0.9
python-docx>=1.1.0
pandas>=2.0.0
openpyxl>=3.1.0
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-core>=0.3.0
langchain-text-splitters>=0.3.0
chromadb>=0.5.0
openai>=1.50.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
pydantic[email]>=2.7.0
python-dotenv>=1.0.0
PyJWT>=2.8.0
requests>=2.31.0
numpy>=1.26.0
